Create searchHotkeys.js
call setupSearchHotkeys() in hotkeys.js
